OMEinsumContractionOrders extension
