+++ Project.toml
@@ -24,7 +24,6 @@ NamedGraphs = "678767b0-92e7-4007-89e4-4527a8725b19"
 PackageExtensionCompat = "65ce6f38-6b18-4e1d-a461-8949797d7930"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 SerializedElementArrays = "d3ce8812-9567-47e9-a7b5-65a6d70a3065"
 SimpleTraits = "699a6c99-e7fa-54fc-8d76-47d257e15c1d"
 SparseArrayKit = "a9a3c162-d163-4c15-8926-b8794fbefed2"
@@ -38,10 +37,12 @@ TupleTools = "9d95972d-f1c8-5527-a6e0-b4b365fa01f6"
 [weakdeps]
 EinExprs = "b1794770-133b-4de1-afb4-526377e9f4c5"
 Observers = "338f10d5-c7f1-4033-a7d1-f9dec39bcaa0"
+OMEinsumContractionOrders = "6f22d1fd-8eed-4bb7-9776-e7d684900715"
 
 [extensions]
 ITensorNetworksEinExprsExt = "EinExprs"
 ITensorNetworksObserversExt = "Observers"
+ITensorNetworksOMEinsumContractionOrdersExt = "OMEinsumContractionOrders"
 
 [compat]
 AbstractTrees = "0.4.4"
@@ -62,8 +63,8 @@ KrylovKit = "0.6, 0.7"
 NamedGraphs = "0.5.1"
 NDTensors = "0.3"
 Observers = "0.2"
+OMEinsumContractionOrders = "0.8.3"
 PackageExtensionCompat = "1"
-Requires = "1.3"
 SerializedElementArrays = "0.1"
 SimpleTraits = "0.9"
 SparseArrayKit = "0.3"
@@ -78,6 +79,7 @@ julia = "1.10"
 [extras]
 EinExprs = "b1794770-133b-4de1-afb4-526377e9f4c5"
 Observers = "338f10d5-c7f1-4033-a7d1-f9dec39bcaa0"
+OMEinsumContractionOrders = "6f22d1fd-8eed-4bb7-9776-e7d684900715"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
