+++ src/ITensorNetworks.jl
@@ -66,11 +66,7 @@ include("lib/ModelHamiltonians/src/ModelHamiltonians.jl")
 include("lib/ModelNetworks/src/ModelNetworks.jl")
 
 using PackageExtensionCompat: @require_extensions
-using Requires: @require
 function __init__()
   @require_extensions
-  @require OMEinsumContractionOrders = "6f22d1fd-8eed-4bb7-9776-e7d684900715" include(
-    "requires/omeinsumcontractionorders.jl"
-  )
 end
 end
